drivers: staging: ksu: Allow kprobes hooks optional on GKI
Some apps will detect that KernelSU uses kprobes and think that the device
environment is abnormal. Therefore, to avoid being detected, it is best to
follow the official documentation of KernelSU to manually hook, just like
doing it on a traditional kernel.

This commit allows us to manually hook while keeping `CONFIG_KPROBES` enabled.

Notes:
The kernel having kprobes support does not mean that the environment is
abnormal, because GKI has kprobes support by default, and disabling kprobes
support will make it impossible to start the device.

[TogoFire]: The use of KProbes is not recommended due to security concerns, performance impacts, and for being trash.

Signed-off-by: rsuntk <rissu.ntk@gmail.com>
Change-Id: I52c3ca4404c58403585b1d414348d46464f2d426
Signed-off-by: TogoFire <togofire@mailfence.com>
Co-Authored-By: Pzqqt <821026875@qq.com>